Refactor action composable code
