--- app/src/main/java/protect/card_locker/compose/OverflowMenuEntry.kt
@@ -1,5 +1,0 @@
-package protect.card_locker.compose
-
-enum class OverflowMenuEntry {
-    IMAGE_GALLERY
-}
