--- app/src/main/java/protect/card_locker/compose/OverflowMenuParameter.kt
@@ -1,8 +1,0 @@
-package protect.card_locker.compose
-
-enum class OverflowMenuParameter {
-
-    //IMAGE_GALLERY entry parameters
-    LOYALTY_CARD_ID,
-    IMAGE_LOCATION_TYPE
-}
